+++ app/Http/Controllers/TaskController.php
@@ -4,14 +4,34 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
-use App\Models\Task;
+use App\Models\{
+    Task,
+    Phase,
+    User
+};
 
 class TaskController extends Controller
 {
+    public $taskModel, $phaseModel, $userModel;
+
+    public function __construct(
+        Task $taskModel,
+        Phase $phaseModel,
+        User $userModel
+    ) {
+        $this->taskModel = $taskModel;
+        $this->phaseModel = $phaseModel;
+        $this->userModel = $userModel;
+    }
 
     public function kanban()
     {
         return view('tasks.index');
+    }
+
+    public function statistics()
+    {
+        return view('tasks.statistics');
     }
 
     /**
@@ -19,7 +39,7 @@ class TaskController extends Controller
      */
     public function index()
     {
-        return \App\Models\Phase::with('tasks.user')->get();
+        return $this->phaseModel->with('tasks.user')->get();
     }
 
     /**
@@ -27,7 +47,7 @@ class TaskController extends Controller
      */
     public function users()
     {
-        return \App\Models\User::all();
+        return $this->userModel->all();
     }
 
     /**
@@ -44,7 +64,8 @@ class TaskController extends Controller
     public function store(StoreTaskRequest $request)
     {
         // Create a new task from the $request
-        $task = Task::create($request->validated());
+        $data = $request->validated();
+        $task = $this->taskModel->create($request->validated());
     }
 
     /**
@@ -63,19 +84,61 @@ class TaskController extends Controller
         //
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(UpdateTaskRequest $request, Task $task)
+    public function update(UpdateTaskRequest $request)
     {
-        //
+        $data = $request->validated();
+        $taskId = $data['id'];
+        unset($data['id']);
+
+        // Retrieve the Task model instance by its ID
+        $task = Task::find($taskId);
+
+        if ($task) {
+            // Update the model attributes
+            $task->fill($data);
+            $task->save();
+        } else {
+            // Handle the case when the Task with the provided ID is not found
+            // You can log an error or return a response as needed
+        }
     }
+
 
     /**
      * Remove the specified resource from storage.
      */
     public function destroy(Task $task)
     {
-        Task::destroy($task->id);
+        $this->taskModel->destroy($task->id);
+    }
+
+    public function getStatistics() {
+        $usersWithTasks = $this->userModel
+            ->with(['tasks'])
+            ->get();
+
+        $groupedUsers = $usersWithTasks->map(function ($user) {
+            $dueTasks = [];
+            $inProgressTasks = [];
+            $completedTasks = [];
+
+            foreach ($user->tasks as $task) {
+                if ($task->status === 'due') {
+                    $dueTasks[] = $task;
+                } elseif ($task->status === 'in_progress') {
+                    $inProgressTasks[] = $task;
+                } elseif ($task->status === 'completed') {
+                    $completedTasks[] = $task;
+                }
+            }
+
+            $user->due_tasks = $dueTasks;
+            $user->in_progress_tasks = $inProgressTasks;
+            $user->completed_tasks = $completedTasks;
+
+            return $user;
+        });
+
+        return $groupedUsers;
     }
 }