+++ config/settings.php
@@ -1,0 +1,10 @@
+<?php
+
+return [
+    // Define background colors for task cards based on their status.
+    'task_card_background_color_status_wise' => [
+        'due' => "#EE4B2B",         // Background color for tasks that are due
+        'completed' => "#50ff40",   // Background color for completed tasks
+        'in_progress' => "#FFFFFF"  // Background color for tasks that are in progress
+    ]
+];