+++ database/seeders/PhaseSeeder.php
@@ -14,13 +14,15 @@ class PhaseSeeder extends Seeder
     public function run(): void
     {
         Phase::factory()
-            ->count(5)
+            ->count(6)
             ->sequence(
                 ['name' => 'Backlog'],
                 ['name' => 'To Do'],
                 ['name' => 'Doing'],
                 ['name' => 'Done'],
-                ['name' => 'Archived'],)
+                ['name' => 'Archived'],
+                ['name' => 'completion'],
+                )
             ->create();
     }
 }