+++ resources/views/layouts/navigation.blade.php
@@ -19,6 +19,10 @@
                     <x-nav-link :href="route('tasks.index')" :active="request()->routeIs('tasks.index')">
                         {{ __('Tasks') }}
                     </x-nav-link>
+
+                    <x-nav-link :href="route('tasks.statistics')" :active="request()->routeIs('tasks.statistics')">
+                        {{ __('Statistics') }}
+                    </x-nav-link>
                 </div>
             </div>