+++ routes/api.php
@@ -17,8 +17,14 @@ use Illuminate\Support\Facades\Route;
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
+
+Route::middleware('auth:sanctum')->get('/app-settings', [App\Http\Controllers\GeneralController::class, 'getAppSettings']);
 Route::middleware('auth:sanctum')->get('/tasks', [App\Http\Controllers\TaskController::class, 'index']);
+Route::middleware('auth:sanctum')->get('/tasks', [App\Http\Controllers\TaskController::class, 'index']);
+Route::middleware('auth:sanctum')->get('/users/statistics', [App\Http\Controllers\TaskController::class, 'getStatistics']);
 Route::middleware('auth:sanctum')->post('/tasks', [App\Http\Controllers\TaskController::class, 'store']);
+Route::middleware('auth:sanctum')->patch('/tasks', [App\Http\Controllers\TaskController::class, 'update']);
 Route::middleware('auth:sanctum')->delete('/tasks/{task}', [App\Http\Controllers\TaskController::class, 'destroy']);
 Route::middleware('auth:sanctum')->get('/users', [App\Http\Controllers\TaskController::class, 'users']);
 Route::middleware('auth:sanctum')->get('/phases/{phase}', [App\Http\Controllers\PhaseController::class, 'show']);
+Route::middleware('auth:sanctum')->delete('/phases/{phase}', [App\Http\Controllers\PhaseController::class, 'destroy']);