+++ CHANGES.md
@@ -2,3 +2,4 @@
 
 ## 0.1
 - Initial release
+- npm install sweetalert2 (Sweet alert for pretty popups)